Integrated clang-cl build under windows (#357)
* Integrated clang-cl build with cmake
* Updated cmake.yml workflow action file to test the build
* Fixed compilation of  google test library by disabling warnings when compile with clang-cl
* Fixed the warnings in the source codes of areg sdk

+++ framework/areg/base/GEMacros.h
@@ -54,7 +54,7 @@
  * \brief   Return size of statically defined array
  **/
 #ifndef MACRO_ARRAYLEN
-    #define MACRO_ARRAYLEN(a)   (sizeof(a) / sizeof(a[0]))
+    #define MACRO_ARRAYLEN(a)   static_cast<uint32_t>(sizeof(a) / sizeof(a[0]))
 #endif  // MACRO_ARRAYLEN
 
 /**
