+++ framework/areg/base/TEArrayList.hpp
@@ -933,10 +933,10 @@ inline void TEArrayList< VALUE >::insertAt(uint32_t startAt, const std::vector< 
 
     if (newArray.empty() == false)
     {
-        uint32_t limit = 0;
+        int32_t limit = 0;
         if ((getSize() + newArray.size()) > NECommon::MAX_CONTAINER_SIZE)
         {
-            limit = NECommon::MAX_CONTAINER_SIZE - (getSize() + static_cast<uint32_t>(newArray.size()));
+            limit = static_cast<int>(NECommon::MAX_CONTAINER_SIZE - (getSize() + static_cast<uint32_t>(newArray.size())));
         }
 
         ARRAYPOS cit = getPosition(startAt);
@@ -993,20 +993,20 @@ inline bool TEArrayList< VALUE >::removeElem( const VALUE & elemRemove, uint32_t
 }
 
 template<typename VALUE >
-void TEArrayList< VALUE >::shift(uint32_t startAt, int count)
+void TEArrayList< VALUE >::shift(uint32_t startAt, int  count)
 {
     if ((mValueList.size() != 0) && (startAt < mValueList.size()) && (count != 0))
     {
         if (count > 0)
         {
-            if ((getSize() + count) > NECommon::MAX_CONTAINER_SIZE)
+            if (static_cast<uint32_t>(static_cast<int>(getSize()) + count) > NECommon::MAX_CONTAINER_SIZE)
             {
                 count = static_cast<int>(NECommon::MAX_CONTAINER_SIZE - getSize());
             }
 
             VALUE* values = mValueList.data();
             uint32_t size = static_cast<uint32_t>(mValueList.size());
-            mValueList.resize(size + count);
+            mValueList.resize(static_cast<uint32_t>(static_cast<int>(size) + count));
             NEMemory::moveElems<VALUE>(values + startAt + count, values + startAt, size - startAt);
         }
         else if (startAt != 0)
@@ -1021,7 +1021,7 @@ void TEArrayList< VALUE >::shift(uint32_t startAt, int count)
             }
 
             NEMemory::moveElems<VALUE>(values + startAt - count, values + startAt, size - startAt);
-            mValueList.resize(size - count);
+            mValueList.resize(static_cast<uint32_t>(static_cast<int>(size) - count));
         }
     }
 }
